Bug 116958 - Make breakpad work again. r=fabrice
Our build script utils.sh set CC, CXX, CFLAGS and other env variables
to force cargo use a specific toolchain.  However, it also affects how
it build build-dependencies that should be built with the toolchain
for the host (x86 for example), not for target (ARM for example).

This patch create an empty rust package that |libloading|, the package
in question, is one of its build-dependencies.  The empty package,
native_build_deps, is built separatedly without these settings to make
|libloading| is built with a correct toolchain; aka. compiler and
linker.

Since native_build_deps is an empty package, so building without
forcing cargo to use a specific toolchain doesn't hurt it.