Bug 103969 - Add a blob type to sidl. r=seinlin.li
This patch adds a new 'blob' type to sidl. This type is mapped to the following types:
- In Javascript, the standard Blob type.
- In Rust, a new Blob type wrapping the data and mime type.

The biggest change is the codegen for calls from Javascript since getting the content
of the blob with arrayBuffer() is async, we create a set of promise for each parameter
needing to be resolved, and call the daemon once they are all resolved.